state_tracking/requirements.txt
python-chess>=1.9.0
transformers>=4.30.0
torch>=2.0.0
groq>=0.3.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
colorama>=0.4.6
pytest>=7.4.0
matplotlib>=3.7.0
seaborn>=0.12.0 